refactor(seed) : File Complexity Seed Seperated
